Refactor import paths for better modularity
Updated import paths in test_mediator_design_pattern.py, crossref.py, and pdf_manipulation.py for improved organization. This change enhances code readability and maintainability by relocating modules under a more consistent directory structure.
